requirements.txt
fastapi
uvicorn
python-dotenv
pdfplumber
requests
google-genai
pandas
openpyxl
python-multipart
